requirements.txt
Flask==3.0.2
groq
python-dotenv==1.0.1
requests==2.31.0
PyJWT[crypto]==2.8.0
flask-cors